+++ src/perception/lidar_road_detectiion/launch/detector.launch
@@ -1,0 +1,4 @@
+<launch>
+    <node name="lidar_road_detectiion" pkg="lidar_road_detectiion" type="main.py" output="screen">
+    </node>
+</launch>
